Moved field setters out of test-specific code
